almost done

+++ includes/out_funcs.h
@@ -8,8 +8,8 @@ void	out_funcs(va_list arg, int *total, t_specifier *specifier);
 void	char_print(char c, int *total, t_specifier *specifier);
 void	str_print(char *str, int *total, t_specifier *specifier);
 void	pointer_print(void *ptr, int *total, t_specifier *specifier);
-void	digit_print(int nbr, int *total, t_specifier *specifier);
-void	u_digit_print(uint32_t nbr, int *total, t_specifier *specifier);
-void	hex_print(uint32_t nbr, int *total, t_specifier *specifier);
+void	digit_print(int64_t nbr, int *total, t_specifier *specifier);
+void	u_digit_print(uint64_t nbr, int *total, t_specifier *specifier);
+void	hex_print(uint64_t nbr, int *total, t_specifier *specifier);
 
 #endif
